+++ src/main/java/com/project/draw/domain/auth/domain/RefreshToken.java
@@ -12,7 +12,7 @@ import org.springframework.data.redis.core.index.Indexed;
 @RedisHash
 public class RefreshToken {
     @Id
-    private String accountId;
+    private String email;
 
     @Indexed
     private String token;
