+++ src/main/java/com/project/draw/domain/auth/domain/repository/RefreshTokenRepository.java
@@ -7,4 +7,5 @@ import java.util.Optional;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
     Optional<RefreshToken> findByToken(String refreshToken);
+    RefreshToken findByEmail(String email);
 }
