+++ src/main/java/com/project/draw/domain/auth/presentation/dto/response/TokenResponse.java
@@ -1,0 +1,20 @@
+package com.project.draw.domain.auth.presentation.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class TokenResponse {
+
+    private final String accessToken;
+
+    private final LocalDateTime expiredAt;
+
+    private final String refreshToken;
+
+}
